feat: create room and add air-conditioners to them

+++ src/controllers/air-conditioner.controller.ts
@@ -2,7 +2,6 @@ import { type Request, type Response } from 'express'
 import { type AirConditionerService } from '@/services'
 import { ResourceNotFound } from '@/errors'
 import { idSchema } from '@/schemas'
-import { ZodError } from 'zod'
 
 export class AirConditionerController {
   constructor (private readonly airConditionerService: AirConditionerService) {}
@@ -13,30 +12,40 @@ export class AirConditionerController {
   }
 
   getAirConditionerById = async (req: Request, res: Response) => {
+    const parsedParams = idSchema.safeParse(req.params)
+
+    if (!parsedParams.success) {
+      return res.status(400).json(
+        {
+          message: 'Validation error', error: parsedParams.error.format()
+        }
+      )
+    }
+
+    const { id } = parsedParams.data
+
     try {
-      const { id } = idSchema.parse(req.params)
       const airConditioner = await this.airConditionerService.getAirConditionerById(id)
       res.json(airConditioner)
     } catch (error) {
       if (error instanceof ResourceNotFound) {
         res.status(404).json({ error: error.message })
-      } else if (error instanceof ZodError) {
-        res.status(400).json({ message: 'Validation error', error: error.format() })
       }
     }
   }
 
-  createAirConditioner = async (req: Request, res: Response) => {
-    const airConditioner = req.validatedData
-    const newAirConditioner = await this.airConditionerService.createAirConditioner(
-      airConditioner
-    )
-    res.status(201).json(newAirConditioner)
-  }
-
   updateAirConditioner = async (req: Request, res: Response) => {
-    const id = Number(req.params.id)
-    const airConditioner = req.validatedData
+    const parsedParams = idSchema.safeParse(req.params)
+
+    if (!parsedParams.success) {
+      return res.status(400).json(
+        {
+          message: 'Validation error', error: parsedParams.error.format()
+        }
+      )
+    }
+
+    const { id } = parsedParams.data
 
     try {
       const updatedAirConditioner = await this.airConditionerService.updateAirConditioner(id, airConditioner)
@@ -49,7 +58,17 @@ export class AirConditionerController {
   }
 
   deleteAirConditioner = async (req: Request, res: Response) => {
-    const id = Number(req.params.id)
+    const parsedParams = idSchema.safeParse(req.params)
+
+    if (!parsedParams.success) {
+      return res.status(400).json(
+        {
+          message: 'Validation error', error: parsedParams.error.format()
+        }
+      )
+    }
+
+    const { id } = parsedParams.data
 
     try {
       const deletedAirConditioner = await this.airConditionerService.deleteAirConditioner(id)
@@ -62,7 +81,17 @@ export class AirConditionerController {
   }
 
   updateAirConditionerState = async (req: Request, res: Response) => {
-    const id = Number(req.params.id)
+    const parsedParams = idSchema.safeParse(req.params)
+
+    if (!parsedParams.success) {
+      return res.status(400).json(
+        {
+          message: 'Validation error', error: parsedParams.error.format()
+        }
+      )
+    }
+
+    const { id } = parsedParams.data
     const { state } = req.validatedData
 
     try {
